Add files via upload
The bill text file is used to print the bill in a certain execution of code. It is empty when you run the code again. The cafeteria text file records all the orders for each and every execution of the code. The CHECK text file contains the valid user IDs, you may change the valid IDs as per your need.